refactor: move pi4j context builder into `CrowPiPlatform`
